chore: remove esm package
