+++ packages/cluster/index.d.cts
@@ -1,0 +1,15 @@
+import type { RequestListener } from 'http';
+
+declare namespace cluster {
+	export interface ClusterController {
+		listen(port: number): void;
+	}
+}
+
+declare function cluster(
+	app: RequestListener | {
+		listen: Function
+	}
+): cluster.ClusterController;
+
+export = cluster;
