+++ packages/cluster/index.d.ts
@@ -1,15 +1,7 @@
 import type { RequestListener } from 'http';
 
-declare namespace cluster {
-	export interface ClusterController {
-		listen(port: number): void;
-	}
+export interface ClusterController {
+	listen(port: number): void;
 }
 
-declare function cluster(
-	app: RequestListener | {
-		listen: Function
-	}
-): cluster.ClusterController;
-
-export = cluster;
+export default function (app: RequestListener | { listen: Function }): ClusterController;
