+++ packages/parse/package.json
@@ -3,19 +3,20 @@
   "name": "@polka/parse",
   "repository": "lukeed/polka",
   "description": "WIP",
-  "module": "build.mjs",
+  "type": "module",
+  "module": "build.js",
   "types": "index.d.ts",
-  "main": "build.js",
+  "main": "build.cjs",
   "license": "MIT",
   "exports": {
     ".": {
       "import": {
         "types": "./index.d.ts",
-        "default": "./build.mjs"
+        "default": "./build.js"
       },
       "require": {
         "types": "./index.d.ts",
-        "default": "./build.js"
+        "default": "./build.cjs"
       }
     },
     "./package.json": "./package.json"
