+++ packages/redirect/index.d.cts
@@ -1,0 +1,5 @@
+import type { ServerResponse } from 'http';
+
+declare function redirect(res: ServerResponse, code?: number, location?: string): void;
+
+export = redirect;
