--- packages/redirect/index.d.mts
@@ -1,2 +1,0 @@
-import type { ServerResponse } from 'http';
-export default function (res: ServerResponse, code?: number, location?: string): void;
