+++ packages/send/index.d.ts
@@ -1,14 +1,3 @@
 import type { ServerResponse } from 'http';
-
-declare namespace send {
-	export type OutgoingHeaders = Record<string, string|string[]>;
-}
-
-declare function send(
-	res: ServerResponse,
-	status?: number,
-	data?: any,
-	headers?: send.OutgoingHeaders
-): void;
-
-export = send;
+export type OutgoingHeaders = Record<string, string|string[]>;
+export default function (res: ServerResponse, status?: number, data?: any, headers?: OutgoingHeaders): void;
