+++ packages/send/package.json
@@ -3,19 +3,20 @@
   "name": "@polka/send",
   "repository": "lukeed/polka",
   "description": "A response helper that detects & handles Content-Types",
-  "module": "build.mjs",
-  "types": "index.d.ts",
-  "main": "build.js",
+  "type": "module",
+  "module": "build.js",
+  "types": "index.d.cts",
+  "main": "build.cjs",
   "license": "MIT",
   "exports": {
     ".": {
       "import": {
-        "types": "./index.d.mts",
-        "default": "./build.mjs"
-      },
-      "require": {
         "types": "./index.d.ts",
         "default": "./build.js"
+      },
+      "require": {
+        "types": "./index.d.cts",
+        "default": "./build.cjs"
       }
     },
     "./package.json": "./package.json"
